requirements.txt
SPARQLWrapper==2.0.0
rdflib==7.0.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.8.3
PyJWT==2.8.0
flask-mail==0.9.1

